+++ src/main/java/com/zju/vis/print_backend/service/RawMaterialService.java
@@ -293,9 +293,14 @@ public class RawMaterialService {
     public ResultVo addRawMaterial(RawMaterialStandardVo rawMaterialStandard) {
         DeStandardizeResult result = deStandardizeRawMaterial(rawMaterialStandard);
         RawMaterial rawMaterial = result.getRawMaterial();
-        if(rawMaterialRepository.findRawMaterialByRawMaterialIndex(rawMaterial.getRawMaterialIndex()) != null){
-            return ResultVoUtil.error("原料编码重复");
+        if(rawMaterialRepository.findRawMaterialByRawMaterialName(rawMaterial.getRawMaterialName()) != null){
+            return ResultVoUtil.error("原料名称重复");
         }
+        // 1.编号不为空 2.编号与数据库中已有编号重复
+        // if(!utils.isEmptyString(rawMaterial.getRawMaterialIndex()) &&
+        //         rawMaterialRepository.findRawMaterialByRawMaterialIndex(rawMaterial.getRawMaterialIndex()) != null){
+        //     return ResultVoUtil.error("原料编号重复");
+        // }
         List<RelDateRawMaterial> relDateRawMaterialList = result.getRelDateRawMaterialList();
 
         rawMaterial.setRawMaterialId(new Long(0));
@@ -308,30 +313,37 @@ public class RawMaterialService {
 
     //改
     //-------------------------------------------------------------------------
-    public String updateRawMaterial(RawMaterialStandardVo updatedRawMaterial) {
+    public ResultVo updateRawMaterial(RawMaterialStandardVo updatedRawMaterial) {
         RawMaterial originRawMaterial = rawMaterialRepository.findRawMaterialByRawMaterialId(updatedRawMaterial.getRawMaterialId());
         if(originRawMaterial == null){
             ResultVo<RawMaterialStandardVo> result = addRawMaterial(updatedRawMaterial);
-            if(result.checkSuccess()){
-                return "数据库中不存在对应数据,已添加Id为" + result.getData().getRawMaterialId() + "的条目" ;
-            }
-            else{
-                return "原料编码重复";
-            }
+            return result;
         }
+        DeStandardizeResult result = deStandardizeRawMaterial(updatedRawMaterial);
+        RawMaterial rawMaterial = result.getRawMaterial();
+        // 1.新名字与原名字不一致 2.新名字与数据库中已有的名字重复
+        if(!rawMaterial.getRawMaterialName().equals(originRawMaterial.getRawMaterialName()) &&
+                rawMaterialRepository.findRawMaterialByRawMaterialName(rawMaterial.getRawMaterialName()) != null){
+            return ResultVoUtil.error("原料名称重复");
+        }
+        // 1.新编号不为空 2.新编号与原编号不一致 3.新编号与数据库中已有编号重复
+        // if(!utils.isEmptyString(rawMaterial.getRawMaterialIndex()) &&
+        //         !rawMaterial.getRawMaterialIndex().equals(originRawMaterial.getRawMaterialIndex()) &&
+        //         rawMaterialRepository.findRawMaterialByRawMaterialIndex(rawMaterial.getRawMaterialIndex()) != null){
+        //     return ResultVoUtil.error("原料编号重复");
+        // }
+
         // 删除原先单向时间关系
         for(RelDateRawMaterial relDateRawMaterial: originRawMaterial.getRelDateRawMaterialList()){
             relDateRawMaterialService.delete(relDateRawMaterial);
         }
 
         // 重新添加关系以修改内容
-        DeStandardizeResult result = deStandardizeRawMaterial(updatedRawMaterial);
-        RawMaterial rawMaterial = result.getRawMaterial();
         List<RelDateRawMaterial> relDateRawMaterialList = result.getRelDateRawMaterialList();
 
         RawMaterial savedRawMaterial = rawMaterialRepository.save(rawMaterial);
         saveRelDateRawMaterials(savedRawMaterial,relDateRawMaterialList);
-        return "RawMaterial " + originRawMaterial.getRawMaterialName() + " has been changed";
+        return ResultVoUtil.success(updatedRawMaterial);
     }
 
     //删
